requirements.txt
asgiref==3.7.2
Django==4.1.13
djongo==1.3.6
dnspython==2.4.2
pymongo==4.6.1
sqlparse==0.2.4
typing_extensions==4.8.0
djangorestframework==3.14.0
mysqlclient==2.1.0